refactor(web): extract date-range utils and expand feature barrel exports
- Extract DatePeriod type and getDateRangeForPeriod() from useAssignments hook
  into a dedicated utils/date-range.ts for improved testability and reuse
- Expand assignments barrel with useValidationClosedAssignments,
  useAssignmentDetails, getDateRangeForPeriod, and DatePeriod
- Expand compensations barrel with mutation hooks and error types
  (usePaidCompensations, useUpdateCompensation, COMPENSATION_ERROR_KEYS, etc.)
- Expand exchanges barrel with ExchangeStatus type

These barrel additions enable common/ code to use barrel imports per the
ESLint rule, while feature-to-feature imports continue using direct paths
to preserve code splitting.
