+++ packages/web/src/features/assignments/index.ts
@@ -2,7 +2,15 @@
 export { AssignmentsPage } from './AssignmentsPage'
 
 // Hooks
-export { useAssignments, useUpcomingAssignments, usePastAssignments } from './hooks/useAssignments'
+export {
+  useAssignments,
+  useUpcomingAssignments,
+  usePastAssignments,
+  useValidationClosedAssignments,
+  useAssignmentDetails,
+  getDateRangeForPeriod,
+  type DatePeriod,
+} from './hooks/useAssignments'
 export { useAssignmentActions } from './hooks/useAssignmentActions'
 export { useCalendarAssignments, type CalendarAssignment } from './hooks/useCalendarAssignments'
 export { useCalendarTheme } from './hooks/useCalendarTheme'
