+++ packages/web/src/features/exchanges/index.ts
@@ -7,6 +7,7 @@ export {
   useApplyForExchange,
   useRemoveOwnExchange,
   useAddToExchange,
+  type ExchangeStatus,
 } from './hooks/useExchanges'
 export { useExchangeActions } from './hooks/useExchangeActions'
